:recycle: Separate global train subprocess tests
